backend/functions/send_notifications/requirements.txt
google-cloud-firestore==2.19.0
firebase-admin==6.6.0
functions-framework==3.8.1
